requirements.txt
# Core AI/ML Libraries
vertexai>=1.38.0
google-cloud-aiplatform>=1.34.0
openai>=1.0.0
anthropic>=0.25.0

# Google Cloud Services
google-cloud-vision>=3.4.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0

# GUI Framework
wxPython>=4.2.0

# Web Scraping & HTML Processing
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0

# Image Processing & OCR
opencv-python>=4.8.0
Pillow>=10.0.0
pytesseract>=0.3.10
easyocr>=1.7.0

# Data Processing
numpy>=1.24.0

# File Processing
ebooklib>=0.18
